Fine-tune preprocess canvas so it only rerenders when needed
